+++ src/main/resources/db/changelog/v0/update-swiss-re-feed.sql
@@ -1,0 +1,5 @@
+--liquibase formatted sql
+
+--changeset event-api-migrations:update-swiss-re-feed.sql runOnChange:true
+
+update feeds set providers = '{"tornado.canada-gov", "tornado.australian-bm", "tornado.osm-wiki", "tornado.noaa", "wildfire.frap.cal", "tornado.des-inventar-sendai"}' where alias = 'swissre-02';