+++ src/test/java/io/kontur/eventapi/config/WorkerSchedulerTest.java
@@ -5,9 +5,11 @@ import io.kontur.eventapi.gdacs.job.GdacsSearchJob;
 import io.kontur.eventapi.job.EventCombinationJob;
 import io.kontur.eventapi.job.FeedCompositionJob;
 import io.kontur.eventapi.job.NormalizationJob;
+import io.kontur.eventapi.noaatornado.job.NoaaTornadoImportJob;
 import io.kontur.eventapi.pdc.job.HpSrvMagsJob;
 import io.kontur.eventapi.pdc.job.HpSrvSearchJob;
 import io.kontur.eventapi.firms.jobs.FirmsImportJob;
+import io.kontur.eventapi.staticdata.job.StaticImportJob;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
@@ -25,9 +27,11 @@ class WorkerSchedulerTest {
     private final GdacsSearchJob gdacsSearchJob = mock(GdacsSearchJob.class);
     private final FirmsImportJob firmsImportJob = mock(FirmsImportJob.class);
     private final EmDatImportJob emDatImportJob = mock(EmDatImportJob.class);
+    private final StaticImportJob staticImportJob = mock(StaticImportJob.class);
+    private final NoaaTornadoImportJob noaaTornadoImportJob = mock(NoaaTornadoImportJob.class);
 
     private final WorkerScheduler scheduler = new WorkerScheduler(hpSrvSearchJob, hpSrvMagsJob, gdacsSearchJob, normalizationJob, eventCombinationJob,
-            feedCompositionJob, firmsImportJob, emDatImportJob);
+            feedCompositionJob, firmsImportJob, emDatImportJob, staticImportJob, noaaTornadoImportJob);
 
     @AfterEach
     public void resetMocks() {
@@ -35,9 +39,11 @@ class WorkerSchedulerTest {
         Mockito.reset(hpSrvMagsJob);
         Mockito.reset(gdacsSearchJob);
         Mockito.reset(firmsImportJob);
+        Mockito.reset(staticImportJob);
         Mockito.reset(normalizationJob);
         Mockito.reset(eventCombinationJob);
         Mockito.reset(feedCompositionJob);
+        Mockito.reset(noaaTornadoImportJob);
     }
 
     @Test
@@ -88,6 +94,37 @@ class WorkerSchedulerTest {
         verify(firmsImportJob, times(1)).run();
     }
 
+    @Test
+    public void startStaticImportJob() {
+        ReflectionTestUtils.setField(scheduler, "staticImportEnabled", "true");
+        scheduler.startStaticImport();
+
+        verify(staticImportJob, times(1)).run();
+    }
+
+    @Test
+    public void skipStaticImportJob() {
+        ReflectionTestUtils.setField(scheduler, "staticImportEnabled", "false");
+        scheduler.startStaticImport();
+
+        verify(staticImportJob, never()).run();
+    }
+
+    @Test
+    public void startNoaaTornadoImportJob() {
+        ReflectionTestUtils.setField(scheduler, "noaaTornadoImportEnabled", "true");
+        scheduler.startNoaaTornadoImport();
+
+        verify(noaaTornadoImportJob, times(1)).run();
+    }
+
+    @Test
+    public void skipNoaaTornadoImportJob() {
+        ReflectionTestUtils.setField(scheduler, "noaaTornadoImportEnabled", "false");
+        scheduler.startNoaaTornadoImport();
+
+        verify(noaaTornadoImportJob, never()).run();
+    }
 
     @Test
     public void startNormalizationJob() {